+++ packages/kit/src/core/make_package/index.js
@@ -5,6 +5,8 @@ import * as path from 'path';
 import { preprocess } from 'svelte/compiler';
 import { mkdirp, rimraf } from '../filesystem/index.js';
 
+const essential_files = ['README', 'LICENSE', 'CHANGELOG', '.gitignore', '.npmignore'];
+
 /**
  * @param {import('types/config').ValidatedConfig} config
  * @param {string} cwd
@@ -12,8 +14,10 @@ import { mkdirp, rimraf } from '../filesystem/index.js';
 export async function make_package(config, cwd = process.cwd()) {
 	rimraf(path.join(cwd, config.kit.package.dir));
 
-	// Generate type definitions first so hand-written types can overwrite generated ones
-	await emit_dts(config);
+	if (config.kit.package.emitTypes) {
+		// Generate type definitions first so hand-written types can overwrite generated ones
+		await emit_dts(config);
+	}
 
 	const files_filter = create_filter(config.kit.package.files);
 	const exports_filter = create_filter({
@@ -25,27 +29,15 @@ export async function make_package(config, cwd = process.cwd()) {
 
 	const pkg = JSON.parse(fs.readFileSync(path.join(cwd, 'package.json'), 'utf8'));
 
-	const package_pkg = {
-		name: pkg.name,
-		version: pkg.version,
-		description: pkg.description,
-		keywords: pkg.keywords,
-		homepage: pkg.homepage,
-		bugs: pkg.bugs,
-		license: pkg.license,
-		author: pkg.author,
-		contributors: pkg.contributors,
-		funding: pkg.funding,
-		repository: pkg.repository,
-		dependencies: pkg.dependencies,
-		private: pkg.private,
-		publishConfig: pkg.publishConfig,
-		type: 'module',
-		/** @type {Record<string, string>} */
-		exports: {
-			'./package.json': './package.json'
-		}
-	};
+	delete pkg.scripts;
+	pkg.type = 'module'; // type must be 'module'
+
+	const user_defined_exports = 'exports' in pkg;
+
+	// We still want to always predefine some exports
+	// like package.json that is used by other packages
+	if (!pkg.exports) pkg.exports = {};
+	pkg.exports['./package.json'] = './package.json';
 
 	for (const file of files) {
 		if (!files_filter(file)) continue;
@@ -89,28 +81,31 @@ export async function make_package(config, cwd = process.cwd()) {
 
 		write(path.join(cwd, config.kit.package.dir, out_file), out_contents);
 
-		if (exports_filter(file)) {
-			const entry = `./${out_file}`;
-			package_pkg.exports[entry] = entry;
+		if (!user_defined_exports && exports_filter(file)) {
+			const entry = `./${out_file.replace(/\\/g, '/')}`;
+			const key = entry.endsWith('/index.js') ? entry.slice(0, -'/index.js'.length) : entry;
+			pkg.exports[key] = entry;
 		}
 	}
 
-	const main = package_pkg.exports['./index.js'] || package_pkg.exports['./index.svelte'];
+	const main = pkg.exports['./index.js'] || pkg.exports['./index.svelte'];
 
-	if (main) {
-		package_pkg.exports['.'] = main;
+	if (!user_defined_exports && main) {
+		pkg.exports['.'] = main;
 	}
 
-	write(
-		path.join(cwd, config.kit.package.dir, 'package.json'),
-		JSON.stringify(package_pkg, null, '  ')
-	);
+	write(path.join(cwd, config.kit.package.dir, 'package.json'), JSON.stringify(pkg, null, '  '));
 
-	const project_readme = path.join(cwd, 'README.md');
-	const package_readme = path.join(cwd, config.kit.package.dir, 'README.md');
+	const whitelist = fs.readdirSync(cwd).filter((file) => {
+		const lowercased = file.toLowerCase();
+		return essential_files.some((name) => lowercased.startsWith(name.toLowerCase()));
+	});
+	for (const pathname of whitelist) {
+		const full_path = path.join(cwd, pathname);
+		if (fs.lstatSync(full_path).isDirectory()) continue; // just to be sure
 
-	if (fs.existsSync(project_readme) && !fs.existsSync(package_readme)) {
-		fs.copyFileSync(project_readme, package_readme);
+		const package_path = path.join(cwd, config.kit.package.dir, pathname);
+		if (!fs.existsSync(package_path)) fs.copyFileSync(full_path, package_path);
 	}
 }
 
@@ -232,15 +227,23 @@ export async function emit_dts(config) {
 }
 
 async function try_load_svelte2tsx() {
-	try {
-		const svelte2tsx = (await import('svelte2tsx')).emitDts;
-		if (!svelte2tsx) {
-			throw new Error('Old svelte2tsx version');
-		}
-		return svelte2tsx;
-	} catch (e) {
+	const svelte2tsx = await load();
+	const emit_dts = svelte2tsx.emitDts;
+	if (!emit_dts) {
 		throw new Error(
 			'You need to install svelte2tsx >=0.4.1 if you want to generate type definitions'
 		);
+	}
+	return emit_dts;
+
+	async function load() {
+		try {
+			return await import('svelte2tsx');
+		} catch (e) {
+			throw new Error(
+				'You need to install svelte2tsx and typescript if you want to generate type definitions\n' +
+					e
+			);
+		}
 	}
 }