+++ packages/kit/src/core/make_package/test/fixtures/nested/src/lib/index.js
@@ -1,0 +1,2 @@
+export { default as Test } from './Test.svelte';
+export { default as Internal } from './internal/Test.svelte';