+++ packages/kit/types/config.d.ts
@@ -44,11 +44,9 @@ export type Config = {
 		host?: string;
 		hostHeader?: string;
 		hydrate?: boolean;
-		serviceWorker?: {
-			exclude?: string[];
-		};
 		package?: {
 			dir?: string;
+			emitTypes?: boolean;
 			exports?: {
 				include?: string[];
 				exclude?: string[];
@@ -59,8 +57,8 @@ export type Config = {
 			};
 		};
 		paths?: {
-			base?: string;
 			assets?: string;
+			base?: string;
 		};
 		prerender?: {
 			crawl?: boolean;
@@ -69,6 +67,9 @@ export type Config = {
 			pages?: string[];
 		};
 		router?: boolean;
+		serviceWorker?: {
+			exclude?: string[];
+		};
 		ssr?: boolean;
 		target?: string;
 		trailingSlash?: TrailingSlash;
@@ -97,11 +98,9 @@ export type ValidatedConfig = {
 		host: string;
 		hostHeader: string;
 		hydrate: boolean;
-		serviceWorker: {
-			exclude: string[];
-		};
 		package: {
 			dir: string;
+			emitTypes: boolean;
 			exports: {
 				include: string[];
 				exclude: string[];
@@ -112,8 +111,8 @@ export type ValidatedConfig = {
 			};
 		};
 		paths: {
-			base: string;
 			assets: string;
+			base: string;
 		};
 		prerender: {
 			crawl: boolean;
@@ -122,6 +121,9 @@ export type ValidatedConfig = {
 			pages: string[];
 		};
 		router: boolean;
+		serviceWorker: {
+			exclude: string[];
+		};
 		ssr: boolean;
 		target: string;
 		trailingSlash: TrailingSlash;