+++ packages/kit/types/internal.d.ts
@@ -1,8 +1,16 @@
-import { Load } from './page';
-import { Incoming, GetSession, Handle, ServerResponse, ServerFetch } from './hooks';
 import { RequestHandler } from './endpoint';
+import { Headers, Location, ParameterizedBody } from './helper';
+import { GetSession, Handle, ServerResponse, ServerFetch, StrictBody } from './hooks';
+import { Load } from './page';
 
 type PageId = string;
+
+export type Incoming = Omit<Location, 'params'> & {
+	method: string;
+	headers: Headers;
+	rawBody: StrictBody;
+	body?: ParameterizedBody;
+};
 
 export type Logger = {
 	(msg: string): void;