+++ packages/adapter-cloudflare-workers/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-cloudflare-workers
 
+## 1.0.0-next.15
+
+### Patch Changes
+
+- 4720b67: Default body parsing to binary
+
+## 1.0.0-next.14
+
+### Patch Changes
+
+- 7faf52f: Update and consolidate checks for binary body types
+
 ## 1.0.0-next.13
 
 ### Patch Changes